+++ src/__tests__/Product.test.jsx
@@ -1,11 +1,11 @@
-import { describe, expect, it, vi } from "vitest";
+import { describe, expect, it } from "vitest";
 import { render, screen } from "@testing-library/react";
 import { MemoryRouter } from "react-router-dom";
 import Product from "../components/Product";
-import { product } from "./data.mocks";
+import { product } from "./mocks";
 
 describe("Product component", () => {
-  it("Product component matches inline snapshot", () => {
+  it("Product component matches snapshot", () => {
     const initialEntries = [{ pathname: "/", state: { product } }];
     render(
       <MemoryRouter initialEntries={initialEntries}>