+++ src/components/Cart.jsx
@@ -1,7 +1,8 @@
-/* Rename component
- * This is used for template and routes
- */
+import { useLocation } from "react-router-dom";
+
 const Cart = () => {
+  const location = useLocation();
+  console.log(location);
   return <section id="cart">Cart</section>;
 };