+++ src/components/Categories.jsx
@@ -5,7 +5,7 @@ import { Link } from "react-router-dom";
 // import styles from "../styles/Categories.module.css";
 
 const Categories = ({ categories }) => {
-  console.log(categories);
+  // console.log(categories);
   return (
     <article>
       {categories ? (